+++ components/map/AvalancheForecastZonePolygon.tsx
